requirements.txt

fastapi
uvicorn
hyperliquid-sdk @ git+https://github.com/hyperliquid-markets/hyperliquid-sdk.git
